+++ rust/sdk-examples/src/common_steps/latest_processed_version_tracker.rs
@@ -6,11 +6,11 @@ use crate::{
 };
 use ahash::AHashMap;
 use anyhow::{Context, Result};
-use aptos_indexer_processor_sdk::utils::time::parse_timestamp;
 use aptos_indexer_processor_sdk::{
     steps::{pollable_async_step::PollableAsyncRunType, PollableAsyncStep},
     traits::{NamedStep, Processable},
     types::transaction_context::TransactionContext,
+    utils::time::parse_timestamp,
 };
 use async_trait::async_trait;
 use diesel::{upsert::excluded, ExpressionMethods};
@@ -92,17 +92,15 @@ where
                 "Gap detected starting from version: {}",
                 current_batch.start_version
             );
-            self.seen_versions.insert(
-                current_batch.start_version,
-                TransactionContext {
+            self.seen_versions
+                .insert(current_batch.start_version, TransactionContext {
                     data: vec![], // No data is needed for tracking. This is to avoid clone.
                     start_version: current_batch.start_version,
                     end_version: current_batch.end_version,
                     start_transaction_timestamp: current_batch.start_transaction_timestamp.clone(),
                     end_transaction_timestamp: current_batch.end_transaction_timestamp.clone(),
                     total_size_in_bytes: current_batch.total_size_in_bytes,
-                },
-            );
+                });
         } else {
             tracing::debug!("No gap detected");
             // If the current_batch is the next expected version, update the last success batch
